+++ Gat-IJ-Plugin/GAT-Java-Plugin/src/main/java/UI/panes/SettingPanes/alignStackPane.java
@@ -2,6 +2,7 @@ package UI.panes.SettingPanes;
 
 import Features.Core.Params;
 import Features.Tools.AlignStack;
+import Features.Tools.AlignStackBatch;
 import UI.Handlers.Navigator;
 import UI.panes.WorkflowDashboards.AlignStackDashboard;
 import UI.util.InputValidation;
@@ -23,15 +24,28 @@ public class alignStackPane extends JPanel {
 
     // --- UI components ---
     private JTextField tfImagePath;
+    private JTextField tfInputDir;
     private JButton btnBrowseImage;
+    private JButton btnBrowseInput;
     private JButton btnPreviewImage;
 
     private JSpinner spReferenceFrame;
+    private JSpinner spReferenceFrameBatch;
     private JCheckBox cbUseSIFT;
+    private JCheckBox cbUseSIFTBatch;
+
     private JCheckBox cbUseTemplateMatching;
+    private JRadioButton cbUseTemplateMatchingBatch;
+    private JRadioButton cbUseStackRegBatch;
+
     private JTextField tfOutputDir;
+    private JTextField tfOutputDirBatch;
+    private JTextField tfFileExt;
     private JButton btnBrowseOutput;
+    private JButton btnBrowseOutputBatch;
+
     private JCheckBox cbSaveAlignedStack;
+    private JCheckBox cbSaveAlignedStackBatch;
 
     private JButton runBtn;
 
@@ -45,21 +59,67 @@ public class alignStackPane extends JPanel {
 
         // --- Tabs ---
         JTabbedPane tabs = new JTabbedPane();
-        tabs.addTab("Settings", buildSettingsTab());
+        tabs.addTab("Align Stack (Single) Settings", buildSettingsTab());
+        tabs.addTab("Align Stack (Batch) Settings", buildBatchTab());
 
-        alignStackDashboard = new AlignStackDashboard();
-        tabs.addTab("Alignment Dashboard", alignStackDashboard);
 
         add(tabs, BorderLayout.CENTER);
 
         // --- Run button ---
         JPanel actions = new JPanel(new FlowLayout(FlowLayout.RIGHT));
         runBtn = new JButton("Run Alignment");
-        runBtn.addActionListener(e -> onRun(runBtn));
+        runBtn.addActionListener(e -> onRun(runBtn, tabs));
         actions.add(runBtn);
         add(actions, BorderLayout.SOUTH);
 
         loadDefaults();
+    }
+
+    private JPanel buildBatchTab() {
+        JPanel outer = new JPanel(new BorderLayout());
+        JPanel p = new JPanel();
+        p.setLayout(new BoxLayout(p, BoxLayout.Y_AXIS));
+
+        // Input folder
+        tfInputDir = new JTextField(36);
+        btnBrowseInput = new JButton("Browse…");
+        btnBrowseInput.addActionListener(e -> chooseFile(tfInputDir, JFileChooser.DIRECTORIES_ONLY));
+        p.add(box("Input folder with stacks", row(tfInputDir, btnBrowseInput)));
+
+        // Reference frame
+        spReferenceFrameBatch = new JSpinner(new SpinnerNumberModel(1, 1, 1000, 1));
+        p.add(box("Reference frame", spReferenceFrameBatch));
+
+        // Alignment method
+        cbUseSIFTBatch = new JCheckBox("Use SIFT alignment if tissue moves significantly", true);
+        // Create radio buttons for XY alignment
+        JRadioButton rbTemplateMatching = new JRadioButton("Template Matching", true);
+        JRadioButton rbStackReg = new JRadioButton("StackReg");
+
+        ButtonGroup xyGroup = new ButtonGroup();
+        xyGroup.add(rbTemplateMatching);
+        xyGroup.add(rbStackReg);
+
+        cbUseTemplateMatchingBatch = rbTemplateMatching;
+        cbUseStackRegBatch = rbStackReg;
+
+        p.add(box("Alignment method", column(cbUseSIFTBatch, cbUseTemplateMatchingBatch, cbUseStackRegBatch)));
+
+        // File extension
+        tfFileExt = new JTextField(".tif", 6);
+        p.add(box("File extension", tfFileExt));
+
+        // Output dir + save
+        tfOutputDirBatch = new JTextField(36);
+        btnBrowseOutputBatch = new JButton("Browse…");
+        btnBrowseOutputBatch.addActionListener(e -> chooseFile(tfOutputDirBatch, JFileChooser.DIRECTORIES_ONLY));
+        cbSaveAlignedStackBatch = new JCheckBox("Save aligned stacks (recommended)", true);
+        p.add(box("Output location", column(row(tfOutputDirBatch, btnBrowseOutputBatch), cbSaveAlignedStackBatch)));
+
+        JScrollPane scroll = new JScrollPane(p);
+        scroll.getVerticalScrollBar().setUnitIncrement(16);
+        outer.add(scroll, BorderLayout.CENTER);
+        return outer;
     }
 
     private JPanel buildSettingsTab() {
@@ -98,10 +158,15 @@ public class alignStackPane extends JPanel {
     }
 
     // --- Run logic (only sends params to dashboard) ---
-    private void onRun(JButton runBtn) {
+    private void onRun(JButton runBtn, JTabbedPane tabs) {
         runBtn.setEnabled(false);
-
-        boolean ok = InputValidation.validateImageOrShow(this, tfImagePath.getText());
+        int selected = tabs.getSelectedIndex(); // 0 = single, 1 = batch i think
+        boolean ok = true;
+        if (selected == 0) { // Single alignment tab
+            ok = InputValidation.validateImageOrShow(this, tfImagePath.getText());
+        } else if (selected == 1) { // Batch alignment tab
+            ok = validateDirectoryOrShow(this, tfInputDir.getText());
+        }
         if (!ok) {
             runBtn.setEnabled(true);
             return;
@@ -113,24 +178,31 @@ public class alignStackPane extends JPanel {
             @Override
             protected Void doInBackground() {
                 try {
-                    Params params = buildParamsFromUI();
+                    Params params = buildParamsFromUI(selected);
 
-                    // Run alignment
-                    AlignStack aligner = new AlignStack();
-                    AlignStack.AlignResult result = aligner.run(params);
+                    if (selected == 0) {
+                        // Run alignment
+                        AlignStack aligner = new AlignStack();
+                        AlignStack.AlignResult result = aligner.run(params);
 
-                    // Load the aligned image
-                    String outFile = params.outputDir + File.separator +
-                            new File(params.imagePath).getName().replace(".tif", "_aligned.tif");
-                    alignedImage = IJ.openImage(outFile);
+                        // Load the aligned image
+                        String outFile = params.outputDir + File.separator +
+                                new File(params.imagePath).getName().replace(".tif", "_aligned.tif");
+                        alignedImage = IJ.openImage(outFile);
 
-                    // --- Send aligned image and CSV to dashboard ---
-                    SwingUtilities.invokeLater(() -> {
-                        alignStackDashboard.addAlignedStackWithResults(
-                            result.alignedStack,
-                            result.resultCSV
-                        );
-                    });
+                        // --- Send aligned image and CSV to dashboard ---
+                        SwingUtilities.invokeLater(() -> {
+                            alignStackDashboard = new AlignStackDashboard();
+                            tabs.addTab("Alignment Dashboard", alignStackDashboard);
+                            alignStackDashboard.addAlignedStackWithResults(
+                                result.alignedStack,
+                                result.resultCSV
+                            );
+                        });
+                    } else {
+                        runBatchAlignment(params);
+                    }
+                    
 
                 } catch (Throwable ex) {
                     SwingUtilities.invokeLater(() -> JOptionPane.showMessageDialog(
@@ -146,18 +218,47 @@ public class alignStackPane extends JPanel {
         worker.execute();
     }
 
-    private Params buildParamsFromUI() {
+    private Params buildParamsFromUI(int selectedTab) {
         Params p = new Params();
-        p.imagePath = tfImagePath.getText();
-        p.outputDir = tfOutputDir.getText();
-        p.referenceFrame = ((Number) spReferenceFrame.getValue()).intValue();
-        p.useSIFT = cbUseSIFT.isSelected();
-        p.saveAlignedStack = cbSaveAlignedStack.isSelected();
-        p.uiAnchor = SwingUtilities.getWindowAncestor(this);
-        p.useSIFT = cbUseSIFT.isSelected();
-        p.useTemplateMatching = cbUseTemplateMatching.isSelected();
-        p.saveAlignedStack = cbSaveAlignedStack.isSelected();
+
+        if (selectedTab == 0) {
+            p.imagePath = tfImagePath.getText();
+            p.outputDir = tfOutputDir.getText();
+            p.referenceFrame = ((Number) spReferenceFrame.getValue()).intValue();
+            p.useSIFT = cbUseSIFT.isSelected();
+            p.saveAlignedStack = cbSaveAlignedStack.isSelected();
+            p.uiAnchor = SwingUtilities.getWindowAncestor(this);
+            p.useSIFT = cbUseSIFT.isSelected();
+            p.useTemplateMatching = cbUseTemplateMatching.isSelected();
+            p.saveAlignedStack = cbSaveAlignedStack.isSelected();
+        } else {
+            p.inputDir = tfInputDir.getText();
+            p.outputDir = tfOutputDirBatch.getText();
+            p.referenceFrame = ((Number) spReferenceFrameBatch.getValue()).intValue();
+            p.useSIFT = cbUseSIFTBatch.isSelected();
+            p.useTemplateMatching = cbUseTemplateMatchingBatch.isSelected();
+            p.useStackReg = cbUseStackRegBatch.isSelected();
+            p.fileExt = tfFileExt.getText();
+        }
         return p;
+    }
+
+    private void runBatchAlignment(Params params) throws Exception {
+        // Validate input directory
+        boolean ok = validateDirectoryOrShow(this, tfInputDir.getText());
+        if (!ok) return;
+
+        // Build Params from batch tab
+        Params p = params;
+
+        // Run batch alignment
+        AlignStackBatch.runBatch(p);
+
+        // SwingUtilities.invokeLater(() -> {
+        //     alignStackDashboard = new AlignStackDashboard();
+        //     tabs.addTab("Alignment Dashboard", alignStackDashboard);
+        //     IJ.log("Batch alignment finished. Dashboard ready.");
+        // });
     }
 
     // --- Utilities for UI ---
@@ -226,5 +327,17 @@ public class alignStackPane extends JPanel {
         cbUseSIFT.setSelected(false);
         cbSaveAlignedStack.setSelected(true);
         tfOutputDir.setText(System.getProperty("user.home"));
+    }
+
+    public static boolean validateDirectoryOrShow(JComponent parent, String path) {
+        File f = new File(path);
+        if (!f.exists() || !f.isDirectory()) {
+            JOptionPane.showMessageDialog(parent,
+                    "Directory does not exist:\n" + path,
+                    "Invalid directory",
+                    JOptionPane.ERROR_MESSAGE);
+            return false;
+        }
+        return true;
     }
 }
